Removed packages related to Heroku
Switching over from Heroku to Digital Ocean due to the way Heroku
sets up media files. With Heroku, I would also have to set up an
AWS account for a simple website that shouldn't get an insane
amount of traffic. With switching to Digital Ocean, it will be
cheaper in the end, and I will only have to manage one system.